requirements.txt
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
jazzmin==3.0.1
python-dotenv==1.0.1
Pillow==11.0.0
firebase-admin==6.6.0
